Roaming functionality specified in configuration file.
